+++ app/javascript/packs/pages/admin/graphs.js
